requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
alembic
apscheduler
python-telegram-bot>=20.0
httpx
python-dotenv
